+++ __utils__/get-release-text/src/main.ts
@@ -89,6 +89,17 @@ function getChangelogEntry (changelog: string, version: string) {
           </picture>
         </a>
       </td>
+      <tr>
+        <td align="center" valign="middle">
+          <a href="https://uscreen.de/?utm_source=pnpm&utm_medium=release_notes" target="_blank">
+            <picture>
+              <source media="(prefers-color-scheme: light)" srcset="https://pnpm.io/img/users/uscreen.svg" />
+              <source media="(prefers-color-scheme: dark)" srcset="https://pnpm.io/img/users/uscreen_light.svg" />
+              <img src="https://pnpm.io/img/users/uscreen.svg" width="180" />
+            </picture>
+          </a>
+        </td>
+      </tr>
     </tr>
   </tbody>
 </table>
@@ -135,11 +146,31 @@ function getChangelogEntry (changelog: string, version: string) {
     </tr>
     <tr>
       <td align="center" valign="middle">
+        <a href="https://www.thinkmill.com.au/?utm_source=pnpm&utm_medium=release_notes" target="_blank">
+          <picture>
+            <source media="(prefers-color-scheme: light)" srcset="https://pnpm.io/img/users/thinkmill.svg" />
+            <source media="(prefers-color-scheme: dark)" srcset="https://pnpm.io/img/users/thinkmill_light.svg" />
+            <img src="https://pnpm.io/img/users/thinkmill.svg" width="200" />
+          </picture>
+        </a>
+      </td>
+      <td align="center" valign="middle">
         <a href="https://devowl.io/?utm_source=pnpm&utm_medium=release_notes" target="_blank">
           <picture>
             <source media="(prefers-color-scheme: light)" srcset="https://pnpm.io/img/users/devowlio.svg" />
             <source media="(prefers-color-scheme: dark)" srcset="https://pnpm.io/img/users/devowlio.svg" />
             <img src="https://pnpm.io/img/users/devowlio.svg" width="200" />
+          </picture>
+        </a>
+      </td>
+    </tr>
+    <tr>
+      <td align="center" valign="middle">
+        <a href="https://macpaw.com/?utm_source=pnpm&utm_medium=release_notes" target="_blank">
+          <picture>
+            <source media="(prefers-color-scheme: light)" srcset="https://pnpm.io/img/users/macpaw.svg" />
+            <source media="(prefers-color-scheme: dark)" srcset="https://pnpm.io/img/users/macpaw_light.svg" />
+            <img src="https://pnpm.io/img/users/thinkmill.svg" width="200" />
           </picture>
         </a>
       </td>