+++ config/config/src/getOptionsFromRootManifest.ts
@@ -1,3 +1,4 @@
+import path from 'path'
 import { PnpmError } from '@pnpm/error'
 import {
   type AllowedDeprecatedVersions,
@@ -7,31 +8,45 @@ import {
 } from '@pnpm/types'
 import mapValues from 'ramda/src/map'
 
-export function getOptionsFromRootManifest (manifest: ProjectManifest): {
+export interface OptionsFromRootManifest {
   allowedDeprecatedVersions?: AllowedDeprecatedVersions
   allowNonAppliedPatches?: boolean
   overrides?: Record<string, string>
   neverBuiltDependencies?: string[]
   onlyBuiltDependencies?: string[]
+  onlyBuiltDependenciesFile?: string
   packageExtensions?: Record<string, PackageExtension>
   patchedDependencies?: Record<string, string>
   peerDependencyRules?: PeerDependencyRules
-} {
+}
+
+export function getOptionsFromRootManifest (manifestDir: string, manifest: ProjectManifest): OptionsFromRootManifest {
   // We read Yarn's resolutions field for compatibility
   // but we really replace the version specs to any other version spec, not only to exact versions,
   // so we cannot call it resolutions
   const overrides = mapValues(
     createVersionReferencesReplacer(manifest),
-    manifest.pnpm?.overrides ?? manifest.resolutions ?? {}
+    {
+      ...manifest.resolutions,
+      ...manifest.pnpm?.overrides,
+    }
   )
   const neverBuiltDependencies = manifest.pnpm?.neverBuiltDependencies
   const onlyBuiltDependencies = manifest.pnpm?.onlyBuiltDependencies
+  const onlyBuiltDependenciesFile = manifest.pnpm?.onlyBuiltDependenciesFile
   const packageExtensions = manifest.pnpm?.packageExtensions
   const peerDependencyRules = manifest.pnpm?.peerDependencyRules
   const allowedDeprecatedVersions = manifest.pnpm?.allowedDeprecatedVersions
   const allowNonAppliedPatches = manifest.pnpm?.allowNonAppliedPatches
-  const patchedDependencies = manifest.pnpm?.patchedDependencies
-  const settings = {
+  let patchedDependencies = manifest.pnpm?.patchedDependencies
+  if (patchedDependencies) {
+    patchedDependencies = { ...patchedDependencies }
+    for (const [dep, patchFile] of Object.entries(patchedDependencies)) {
+      if (path.isAbsolute(patchFile)) continue
+      patchedDependencies[dep] = path.join(manifestDir, patchFile)
+    }
+  }
+  const settings: OptionsFromRootManifest = {
     allowedDeprecatedVersions,
     allowNonAppliedPatches,
     overrides,
@@ -41,8 +56,10 @@ export function getOptionsFromRootManifest (manifest: ProjectManifest): {
     patchedDependencies,
   }
   if (onlyBuiltDependencies) {
-    // @ts-expect-error
-    settings['onlyBuiltDependencies'] = onlyBuiltDependencies
+    settings.onlyBuiltDependencies = onlyBuiltDependencies
+  }
+  if (onlyBuiltDependenciesFile) {
+    settings.onlyBuiltDependenciesFile = path.join(manifestDir, onlyBuiltDependenciesFile)
   }
   return settings
 }