+++ pkg-manager/plugin-commands-installation/test/update/interactive.ts
@@ -100,15 +100,18 @@ test('interactively update', async () => {
       choices: [
         headerChoice,
         {
-          name: chalk`is-negative                                                  1.0.0 ❯ 1.0.{greenBright.bold 1}                 `,
+          message: chalk`is-negative                                                  1.0.0 ❯ 1.0.{greenBright.bold 1}                 `,
           value: 'is-negative',
+          name: 'is-negative',
         },
         {
-          name: chalk`micromatch                                                   3.0.0 ❯ 3.{yellowBright.bold 1.10}                `,
+          message: chalk`micromatch                                                   3.0.0 ❯ 3.{yellowBright.bold 1.10}                `,
           value: 'micromatch',
+          name: 'micromatch',
         },
       ],
-      name: 'dependencies',
+      name: '[dependencies]',
+      message: 'dependencies',
     },
   ])
   expect(prompt).toBeCalledWith(
@@ -149,19 +152,23 @@ test('interactively update', async () => {
       choices: [
         headerChoice,
         {
-          name: chalk`is-negative                                                  1.0.1 ❯ {redBright.bold 2.1.0}                 `,
+          message: chalk`is-negative                                                  1.0.1 ❯ {redBright.bold 2.1.0}                 `,
           value: 'is-negative',
+          name: 'is-negative',
         },
         {
-          name: chalk`is-positive                                                  2.0.0 ❯ {redBright.bold 3.1.0}                 `,
+          message: chalk`is-positive                                                  2.0.0 ❯ {redBright.bold 3.1.0}                 `,
           value: 'is-positive',
+          name: 'is-positive',
         },
         {
-          name: chalk`micromatch                                                   3.0.0 ❯ {redBright.bold 4.0.5}                 `,
+          message: chalk`micromatch                                                   3.0.0 ❯ {redBright.bold 4.0.5}                 `,
           value: 'micromatch',
+          name: 'micromatch',
         },
       ],
-      name: 'dependencies',
+      name: '[dependencies]',
+      message: 'dependencies',
     },
   ])
   expect(prompt).toBeCalledWith(
@@ -314,11 +321,13 @@ test('interactively update should ignore dependencies from the ignoreDependencie
             value: '',
           },
           {
-            name: chalk`micromatch                                                   3.0.0 ❯ 3.{yellowBright.bold 1.10}                `,
+            message: chalk`micromatch                                                   3.0.0 ❯ 3.{yellowBright.bold 1.10}                `,
             value: 'micromatch',
+            name: 'micromatch',
           },
         ],
-        name: 'dependencies',
+        name: '[dependencies]',
+        message: 'dependencies',
       },
     ]
   )